add chaos_tests and refactored project

+++ files/control_cluster.sh
@@ -1,11 +1,12 @@
 #! /bin/bash
 set -e
+#set -x
+
+# shellcheck source=/dev/null
+source x.sh
 
 CURRENT_PATH=$(cd $(dirname ${BASH_SOURCE[0]}); pwd)
-PROJECT_PATH=$(dirname "${CURRENT_PATH}")
 BUILD_PATH=${CURRENT_PATH}/build
-BIN_PATH=${CURRENT_PATH}/axiom-ledger
-AXIOM_PROJECT_PATH=/home/hyperchain/code/axiomesh/axiom-ledger
 N=4
 
 
@@ -19,7 +20,7 @@ function stop(){
     lsof_output=$(lsof -i:888${i})
     pid=$(echo "$lsof_output" | awk 'NR==2 {print $2}')
     kill "${pid}"
-    echo "kill node-${i}"
+    echo "kill node${i}"
   done
 }
 
@@ -33,36 +34,39 @@ function status(){
     lsof_output=$(lsof -i:888${i})
     pid=$(echo "$lsof_output" | awk 'NR==2 {print $2}')
     if [[ ! "${pid}" == "" ]]; then
-      echo "${app_name} is running,node:${i}, pid: ${pid}"
+      echo "axiom-ledger is running,node:${i}, pid: ${pid}"
     else
-      echo "${app_name} is stopped"
+      echo "axiom-ledger is stopped"
     fi
   done
 }
 
 function start_by_nohup() {
   echo "===> Staring cluster"
-  for ((i = 0; i < N; i = i + 1)); do
-    ${BUILD_PATH}/node$(($i + 1))/start.sh
+  for ((i = 1; i < N + 1; i = i + 1)); do
+    ${BUILD_PATH}/node${i}/start.sh
   done
 }
 
 function prepare() {
   echo "===> Generating $N nodes configuration"
-  rm -rf "${BUILD_PATH}"
-  mkdir "${BUILD_PATH}"
+  if [ -d "${BUILD_PATH}" ];then
+    rm -rf "${BUILD_PATH}"
+  else 
+    mkdir -p "${BUILD_PATH}"
+  fi
   for ((i = 1; i < N + 1; i = i + 1)); do
-    rm -rf ~/.axiom-ledger/*
     root=${BUILD_PATH}/node${i}
-    mkdir ${root}
-    cp -rf ${AXIOM_PROJECT_PATH}/scripts/package/* ${root}/
-    cp -f /home/hyperchain/go/bin/axiom-ledger ${root}/tools/bin/
-    echo "export AXIOM_LEDGER_PORT_JSONRPC=888${i}" >> ${root}/.env.sh
-    echo "export AXIOM_LEDGER_PORT_WEBSOCKET=999${i}" >> ${root}/.env.sh
-    echo "export AXIOM_LEDGER_PORT_P2P=400${i}" >> ${root}/.env.sh
-    echo "export AXIOM_LEDGER_PORT_PPROF=5312${i}" >> ${root}/.env.sh
-    echo "export AXIOM_LEDGER_PORT_MONITOR=4001${i}" >> ${root}/.env.sh
-    ${root}/axiom-ledger config generate --default-node-index ${i}
+    mkdir -p "$root"
+    cp -rf "$AXIOM_LEDGER_PROJECT_PATH"/scripts/package/* "$root"/
+    cp -f "$AXIOM_LEDGER_BINARY_PATH" "$root"/
+    cp -f "$root"/axiom-ledger "$root"/tools/bin/
+    "$root"/axiom-ledger --repo "$root" config generate --default-node-index ${i}
+    echo "export AXIOM_LEDGER_PORT_JSONRPC=888${i}" >> "$root"/.env.sh
+    echo "export AXIOM_LEDGER_PORT_WEBSOCKET=999${i}" >> "$root"/.env.sh
+    echo "export AXIOM_LEDGER_PORT_P2P=400${i}" >> "$root"/.env.sh
+    echo "export AXIOM_LEDGER_PORT_PPROF=5312${i}" >> "$root"/.env.sh
+    echo "export AXIOM_LEDGER_PORT_MONITOR=4001${i}" >> "$root"/.env.sh
   done
 }
